requirements.txt
jupyter==1.0.0
nltk==3.6.2
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.1
scikit-learn==0.24.2
tensorflow==2.5.0
matplotlib==3.4.2
